+++ R/sjPlotFrequencies.R
@@ -761,7 +761,7 @@ sjp.frq <- function(varCount,
   # ---------------------------------------------------------
   # Check whether ggplot object should be returned or plotted
   # ---------------------------------------------------------
-  if (printPlot) plot(baseplot)
+  if (printPlot) graphics::plot(baseplot)
   # -------------------------------------
   # return results
   # -------------------------------------